requirements.txt
pandas==2.3.3
scikit-learn==1.8.0
matplotlib==3.10.8
seaborn==0.13.2
pcre2==0.6.0
numpy==2.3.5
wandb==0.23.1
torch-geometric==2.7.0